Life

Description

A console implementation of Conway's Game of Life in Haskell

Documentation

type Cell = (Int, Int)

Cell type is given by a 2-tuple of integers (x, y)

type Board = [Cell]

Board is the current world of the game, given as a list of living cells in a 20x20 world

determineNeighbors

Arguments

:: Board
The current board

-> [(Cell, Int)]
A list of 2-tuples containing each living cell and its neighbor count

given a board, returns a list of 2-tuples made of each living cell and its number of neighbors

step

Arguments

:: Board
The current board

-> Board
The next board

step returns the next iteration of the given board

display :: Board -> IO ()

displays a command-line representation of the 20x20 world of cells by mapping over the range and domain [-10..10] checking each cell for membership in the Board (list of living cells)

go :: Board -> Int -> IO ()

main :: IO ()